rename SPIClassInstantiationException to ServiceLoaderException
